requirements.txt
click==8.1.7
Flask==3.1.0
fonttools==4.55.3
Jinja2==3.1.4
matplotlib==3.10.0
python-dotenv==0.19.0
Werkzeug==3.1.3
MarkupSafe==3.0.2

